Move OrganizationsSearch to its own file
