+++ src/pages/Contributors/OrganizationSearch.js
@@ -1,0 +1,112 @@
+import React, { useState } from 'react'
+import { makeStyles } from "@material-ui/core";
+import Box from '@material-ui/core/Box';
+import Grid from '@material-ui/core/Grid';
+import IconButton from '@material-ui/core/IconButton';
+import TextField from "@material-ui/core/TextField";
+import Autocomplete from "@material-ui/lab/Autocomplete";
+import SearchRoundedIcon from '@material-ui/icons/SearchRounded';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+  },
+  autocomplete: {
+  },
+  icon: {
+    backgroundColor: theme.palette.secondary.main,
+    borderBottomLeftRadius: 0,
+    borderBottomRightRadius: 4,
+    borderTopLeftRadius: 0,
+    borderTopRightRadius: 4,
+    color: theme.palette.text.secondary,
+    height: 64,
+  },
+  input: {
+    "& .MuiOutlinedInput-root": {
+      borderBottomLeftRadius: 4,
+      borderBottomRightRadius: 0,
+      borderTopLeftRadius: 4,
+      borderTopRightRadius: 0,
+      height: 64,
+    },
+  },
+}));
+
+/* Autocomplete `inputValue` is text displayed, `value` is selected option */
+const OrganizationSearch = (props) => {
+  const {
+    inputPlaceholder,
+    inputValue: topLevelInputValue,
+    options,
+    setInputValue: setTopLevelInputValue,
+  } = props;
+  const classes = useStyles();
+  const [value, setValue] = useState(null);
+  const [inputValue, setInputValue] = useState(topLevelInputValue);
+
+  const handleClick = (event) => {
+    console.log(event.target);
+  };
+
+  const handleInputChange = (event, value, reason) => {
+    setInputValue(value);
+    setTopLevelInputValue(value);
+  };
+
+  const handleChange = (event, value, reason) => {
+    if (reason === 'select-option') {
+      setValue(value);
+    } else if (reason === 'clear') {
+      setInputValue('');
+      setValue(null);
+    } else {
+      console.log('handleChange value:', value);
+      console.log('handleChange reason:', reason);
+    }
+  };
+
+  return (
+    <Grid container>
+      <Grid item xs={1} />
+      <Grid item xs={10}>
+        <Box display='flex' alignItems='center' className={classes.root}>
+          <Autocomplete
+            className={classes.autocomplete}
+            forcePopupIcon={false}
+            freeSolo
+            fullWidth
+            handleHomeEndKeys
+            inputValue={inputValue}
+            onChange={handleChange}
+            onInputChange={handleInputChange}
+            options={options}
+            renderInput={(params) => (
+              <TextField
+                {...params}
+                className={classes.input}
+                InputProps={{
+                  ...params.InputProps,
+                  type: 'search',
+                }}
+                placeholder={inputPlaceholder}
+              />
+            )}
+            selectOnFocus
+            value={value}
+          />
+          <IconButton
+            type='submit'
+            onClick={handleClick}
+            className={classes.icon}
+            aria-label='search'
+          >
+            <SearchRoundedIcon fontSize='large' />
+          </IconButton>
+        </Box>
+      </Grid>
+      <Grid item xs={1} />
+    </Grid>
+  );
+};
+
+export default OrganizationSearch;
